requirements.txt
requests
Pillow
RPi.GPIO
Flask
Flask-Compress
gevent
